teste_c/processing/requirements.txt
python-dotenv==1.0.1
websockets==12.0
pymongo==4.6.2
redis==5.0.2
